moving iterateTiles back to tileengine.cpp, implenent successfulHack and move this to tileEngine
