+++ src/Battlescape/TileEngine.cpp
@@ -47,6 +47,7 @@
 #include "../Engine/Options.h"
 #include "ProjectileFlyBState.h"
 #include "MeleeAttackBState.h"
+#include "../Savegame/BattleObject.h"
 #include "../fmath.h"
 
 namespace OpenXcom
@@ -163,6 +164,32 @@ bool calculateLineHelper(const Position& origin, const Position& target, FuncNew
 	}
 	return false;
 }
+
+template<typename TileFunc>
+void iterateTiles(SavedBattleGame* save, MapSubset gs, TileFunc func)
+{
+	const auto totalSizeX = save->getMapSizeX();
+	const auto totalSizeY = save->getMapSizeY();
+	const auto totalSizeZ = save->getMapSizeZ();
+
+	gs = MapSubset::intersection(gs, MapSubset{ totalSizeX, totalSizeY });
+	if (gs)
+	{
+		for (int z = 0; z < totalSizeZ; ++z)
+		{
+			auto rowStart = save->getTile(Position{ gs.beg_x, gs.beg_y, z });
+			for (auto stepsY = gs.size_y(); stepsY != 0; --stepsY, rowStart += totalSizeX)
+			{
+				auto curr = rowStart;
+				for (auto stepX = gs.size_x(); stepX != 0; --stepX, curr += 1)
+				{
+					func(curr);
+				}
+			}
+		}
+	}
+}
+
 
 template<typename FuncNewPosition>
 bool calculateParabolaHelper(const Position& origin, const Position& target, double curvature, const Position& delta, FuncNewPosition posFunc)
@@ -5298,6 +5325,68 @@ void TileEngine::updateGameStateAfterScript(BattleActionAttack battleActionAttac
 		calculateLighting(LL_ITEMS, pos, 2, true);
 		calculateFOV(pos, 1, false);
 	}
+}
+
+void TileEngine::successfulHack(SavedBattleGame* _save, Position pos)
+{
+	
+		//for test
+		if (_save->getTile(pos)->getBattleObject())
+		{
+
+			BattleObject* battleObject = _save->getTile(pos)->getBattleObject();
+
+			if (battleObject->getHackingDefence() > 0)
+			{
+				const int doorMCD = battleObject->getRules()->getAlterationMCDNumber();
+				const int radius = battleObject->getRules()->getAlterationMCDRadius();
+
+				Tile* tile = _save->getTile(pos);
+				bool objective = false;
+				Tile* tiles[9];
+
+				static const TilePart parts[8] = { O_WESTWALL,O_NORTHWALL,O_FLOOR,O_WESTWALL,O_NORTHWALL,O_OBJECT,O_OBJECT,O_OBJECT };
+				Position pos = tile->getPosition();
+				MapSubset gs = { std::make_pair(pos.x - radius, pos.x + radius + 1), std::make_pair(pos.y - radius, pos.y + radius + 1) };
+
+				iterateTiles(_save, gs, [&](Tile* tile)
+					{
+						tiles[0] = _save->getTile(Position(pos.x + 1, pos.y, pos.z)); //east wall
+						tiles[1] = _save->getTile(Position(pos.x, pos.y + 1, pos.z)); //south wall
+						tiles[2] = tiles[3] = tiles[4] = tiles[5] = tile;
+						tiles[6] = _save->getTile(Position(pos.x, pos.y - 1, pos.z)); //north bigwall
+						tiles[7] = _save->getTile(Position(pos.x - 1, pos.y, pos.z)); //west bigwall
+
+
+						for (int i = 7; i >= 0; --i)
+						{
+							if (!tiles[i] || !tiles[i]->getMapData(parts[i]))
+								continue; //skip out of map and emptiness
+
+							TilePart currentpart = parts[i], currentpart2;
+
+							int diemcd = tiles[i]->getMapData(currentpart)->getDieMCD();
+							int altmcd = tiles[i]->getMapData(currentpart)->getAltMCD();
+							if (altmcd == doorMCD)
+							{
+								if (diemcd != 0)
+									currentpart2 = tiles[i]->getMapData(currentpart)->getDataset()->getObject(diemcd)->getObjectType();
+								else
+									currentpart2 = currentpart;
+
+								tile->SwitchToAltMCD(currentpart);
+							}
+
+
+						}
+					}
+				);
+
+			}
+
+		}
+		//for test end
+	
 }
 
 }
