+++ src/Battlescape/TileEngine.h
@@ -285,30 +285,8 @@ public:
 	/// Update game state after script hook execution.
 	void updateGameStateAfterScript(BattleActionAttack battleActionAttack, Position pos);
 
-	template<typename TileFunc>
-	void iterateTiles(SavedBattleGame* save, MapSubset gs, TileFunc func)
-	{
-		const auto totalSizeX = save->getMapSizeX();
-		const auto totalSizeY = save->getMapSizeY();
-		const auto totalSizeZ = save->getMapSizeZ();
-
-		gs = MapSubset::intersection(gs, MapSubset{ totalSizeX, totalSizeY });
-		if (gs)
-		{
-			for (int z = 0; z < totalSizeZ; ++z)
-			{
-				auto rowStart = save->getTile(Position{ gs.beg_x, gs.beg_y, z });
-				for (auto stepsY = gs.size_y(); stepsY != 0; --stepsY, rowStart += totalSizeX)
-				{
-					auto curr = rowStart;
-					for (auto stepX = gs.size_x(); stepX != 0; --stepX, curr += 1)
-					{
-						func(curr);
-					}
-				}
-			}
-		}
-	}
+	//Called when hacking was succesfull, and electrical doors may be opened
+	void successfulHack(SavedBattleGame* _save, Position pos);
 
 };
